tests: fix AmazonS3Test in native mode
